Move BlendMode to Rendering
