+++ src/KappaDuck.Quack/Video/Imaging/Surface.cs
@@ -3,7 +3,7 @@
 
 using KappaDuck.Quack.Exceptions;
 using KappaDuck.Quack.Geometry;
-using KappaDuck.Quack.Graphics.Drawing;
+using KappaDuck.Quack.Graphics.Rendering;
 using KappaDuck.Quack.Interop.SDL.Primitives;
 using System.Buffers;
 using System.ComponentModel;
